--fix namespace SendEmailConsumer . --fix namespace GetEmailTraceEndpoint . --fix variable name smsTrace to emailTrace in GetEmailTraceEndpoint . --fix Endpoint Name from SendSmsEndpoint to SendEmailEndpoint in Feature Email. --fix namespace SendEmailRequest . --add directory Services and move Service In It.
