+++ src/Notify/Features/Email/Consumers/SendEmailConsumer.cs
@@ -1,4 +1,4 @@
-﻿namespace Notify.Features.Sms.Consumers;
+﻿namespace Notify.Features.Email.Consumers;
 
 public class SendEmailConsumer(InboxService inboxService)
     : InboxConsumer<SendEmailMessage>(inboxService)
