+++ src/Notify/Features/Email/EmailFeatureConfigure.cs
@@ -1,4 +1,6 @@
-﻿namespace Notify.Features.Email;
+﻿using Notify.Features.Email.Services;
+
+namespace Notify.Features.Email;
 
 public static class EmailFeatureConfigure
 {
