+++ src/Notify/Features/Email/REST/SendEmail/SendEmailRequest.cs
@@ -1,3 +1,3 @@
-﻿namespace Notify.Features.Email.REST.SendSms;
+﻿namespace Notify.Features.Email.REST.SendEmail;
 
 public record SendEmailRequest(Guid MessageId, string To, string Subject, string Body);
