+++ src/Notify/Features/Email/Services/EmailService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net;
 using System.Net.Mail;
 
-namespace Notify.Features.Email;
+namespace Notify.Features.Email.Services;
 
 public class EmailService(
     EmailDbContext dbContext,
