+++ src/Notify/Features/Sms/SmsFeatureConfigure.cs
@@ -1,4 +1,6 @@
-﻿namespace Notify.Features.Sms;
+﻿using Notify.Features.Sms.Services;
+
+namespace Notify.Features.Sms;
 
 public static class SmsFeatureConfigure
 {
